Move create test spectra
