+++ tests/test_train_ms2deepscore.py
@@ -8,9 +8,7 @@ from ms2deepscore.models.SiameseSpectralModel import SiameseSpectralModel
 from ms2deepscore.MS2DeepScore import MS2DeepScore
 from ms2deepscore.SettingsMS2Deepscore import SettingsMS2Deepscore
 from ms2deepscore.train_new_model.train_ms2deepscore import train_ms2ds_model, plot_history
-from tests.create_test_spectra import pesticides_test_spectra
-from tests.test_data_generators import create_test_spectra
-
+from tests.create_test_spectra import pesticides_test_spectra, create_test_spectra
 
 TEST_RESOURCES_PATH = Path(__file__).parent / 'resources'
